# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all recipes",
                "operationId": "8b9e556d4d747bd60232961951494bfb",
                "responses": {
                    "200": {
                        "description": "List of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Spaghetti Carbonara"
                                            },
                                            "unique_code": {
                                                "type": "string",
                                                "example": "REC123"
                                            },
                                            "chef_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "ingredients": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "flour",
                                                    "sugar"
                                                ]
                                            },
                                            "likes_count": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "dislikes_count": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-07-10T12:00:00.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-07-10T12:00:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recipes/create": {
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Store a new recipe",
                "operationId": "f9c70a74d8ce6b92ed00904f1285a929",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "unique_code",
                                    "chef_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Recipe"
                                    },
                                    "unique_code": {
                                        "type": "string",
                                        "example": "REC123"
                                    },
                                    "chef_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "flour",
                                            "sugar"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New Recipe"
                                        },
                                        "unique_code": {
                                            "type": "string",
                                            "example": "REC123"
                                        },
                                        "chef_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "ingredients": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "flour",
                                                "sugar"
                                            ]
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "dislikes_count": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-07-10T12:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-07-10T12:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/recipes/{recipe}/add-to-collection": {
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Add a recipe to a collection",
                "operationId": "5510ee74c07a58ae3c316d4958cea13e",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "collection_id"
                                ],
                                "properties": {
                                    "collection_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe added to collection successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe added to collection successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/recipes/{id}/like": {
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Like a recipe",
                "operationId": "e6962acf586d8e2f2448ab74515041ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe liked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}